Refactor financial components to support scaled integer amounts
- Updated Overview and OverviewGroupTree components to pass symbols for currency handling.
- Modified StockLedger to utilize trading currency from symbols and adjusted cash handling to use scaled integers.
- Enhanced chart components to display values in the correct trading currency.
- Refactored other lines handling to accommodate new currency scaling logic.
- Introduced scale.js for safe conversions between scaled integers and decimal strings.
- Updated format.js to handle currency and symbol scales for consistent display formatting.
- Adjusted grouping and matching logic to use trading currency from symbols.
- Ensured all calculations in stockMath use exact integer arithmetic to prevent float inaccuracies.
- Removed hardcoded currency lists and replaced with dynamic scaling based on fetched data.

+++ backend/src/Controller/MatchController.php
@@ -36,6 +36,6 @@ class MatchController
             return new JsonResponse(['error' => 'mode must be "mirrored" or "direct"'], 400);
         }
 
-        return new JsonResponse($this->matching->findCandidates($currency, (float) $amount, $date, $excludeAccountIds, $mode));
+        return new JsonResponse($this->matching->findCandidates($currency, (int) $amount, $date, $excludeAccountIds, $mode));
     }
 }
